requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.6.0
python-multipart==0.0.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-generativeai==0.3.2
pandas==2.2.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
slowapi==0.1.9
redis==5.0.8
